requirements.txt
lxml
bs4
feedparser
requests
openai
python-dotenv
pymysql